Lab Final/LabFinal/src/main/java/org/example/CheckPalindrome.java
package org.example;

public class CheckPalindrome {
    public static boolean isPalindrome(String str) {
        if (str == null) {
            return false;
        }
        if (str.isEmpty()) {
            return true;
        }
        int left = 0;
        int right = str.length() - 1;
        while (left < right) {
            if (str.charAt(left) != str.charAt(right)) {
                return false;
            }
            left++;
            right--;
        }
        return true;
    }
}
